backend/requirements.txt
pyyaml
sanic
sanic-jwt
stravalib
boto3
pycryptodome
aiobotocore
